Refactored model to have a verifyPassword field, and made related changes in controller and view
