+++ src/main/java/org/launchcode/cheesemvc/controllers/UserController.java
@@ -1,14 +1,13 @@
 package org.launchcode.cheesemvc.controllers;
 
-import org.launchcode.cheesemvc.models.User;
-import org.launchcode.cheesemvc.models.UserData;
+import org.launchcode.cheesemvc.models.user.User;
+import org.launchcode.cheesemvc.models.user.UserData;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.validation.Valid;
 
@@ -30,20 +29,8 @@ public class UserController {
     }
 
     @RequestMapping(value = "add", method = RequestMethod.POST)
-    public String processAddForm(Model model, @ModelAttribute @Valid User user, Errors errors, @RequestParam String verify) {
-        boolean formErrorCheck = false;
-        String pwErrorMessage= "";
-
+    public String processAddForm(Model model, @ModelAttribute @Valid User user, Errors errors) {
         if (errors.hasErrors()) {
-            formErrorCheck = true;
-        }
-        if (!verify.equals(user.getPassword())) {
-            pwErrorMessage = "Passwords did not match";
-            model.addAttribute("pwErrorMessage", pwErrorMessage);
-            model.addAttribute("userEmail", user.getEmail());
-            model.addAttribute("username", user.getUsername());
-        }
-        if (formErrorCheck == true || !pwErrorMessage.isEmpty()) {
             model.addAttribute("title", "Sign Up");
             return "user/add";
         } else {
