+++ src/main/java/org/launchcode/cheesemvc/models/user/User.java
@@ -1,4 +1,4 @@
-package org.launchcode.cheesemvc.models;
+package org.launchcode.cheesemvc.models.user;
 
 import org.hibernate.validator.constraints.Email;
 
@@ -17,9 +17,12 @@ public class User {
     @Size(min=1, max=15, message="Username must be 3-15 characters in length")
     private String username;
 
-    @NotNull
+    @NotNull(message="Password field must not be empty")
     @Size(min=1, max=15, message="Password must be 3-15 characters in length")
     private String password;
+
+    @NotNull(message="Passwords do not match")
+    private String verifyPassword;
 
     @Email (message="Not a valid email address")
     private String email;
@@ -56,6 +59,16 @@ public class User {
 
     public void setPassword(String password) {
         this.password = password;
+        checkPassword();
+    }
+
+    public String getVerifyPassword() {
+        return verifyPassword;
+    }
+
+    public void setVerifyPassword(String verifyPassword) {
+        this.verifyPassword = verifyPassword;
+        checkPassword();
     }
 
     public String getEmail() {
@@ -64,6 +77,10 @@ public class User {
 
     public void setEmail(String email) {
         this.email = email;
+    }
+
+    public LocalDateTime getTimeStamp() {
+        return timeStamp;
     }
 
     @Override
@@ -86,4 +103,13 @@ public class User {
                 ", username='" + username + '\'' +
                 '}';
     }
-}
+
+    private void checkPassword() {
+            if (!this.password.equals(this.verifyPassword)) {
+            this.verifyPassword = null;
+       }
+     }
+
+   }
+
+
